# models/models.go
// Package main provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package models

// Item defines model for Item.
type Item struct {
	Description *string `json:"description,omitempty"`
	Id          *string `json:"id,omitempty"`
	Name        *string `json:"name,omitempty"`
}

// PostItemsJSONRequestBody defines body for PostItems for application/json ContentType.
type PostItemsJSONRequestBody = Item

// PutItemsIdJSONRequestBody defines body for PutItemsId for application/json ContentType.
type PutItemsIdJSONRequestBody = Item